refactor(api-example): enhance session authentication and simplify imports
- Add optional data parameter to Auth method
- Simplify session data update logic
- Update docker-compose configuration
- Correct import paths for local packages
